+++ HISTORY.rst
@@ -2,6 +2,15 @@
 History
 =======
 
+0.2.0 (2019-07-05)
+------------------
+
+* Added game replay support.
+* Added action to register generic services.
+* Added turn action for the troop agents.
+* Added a new map (map_08)
+* Minor bug fixes.
+
 0.1.0 (2019-06-13)
 ------------------